Menu refactored and mobile menu updated
